feat(seo): add dynamic seo metadata and json-ld structured data for nft detail pages
